(#10) Seems that this does not work.
This architecture seems to not work because we are returning
instances of FunctionExpression.
Although the elementary functions returns via clone and deriv pointers the
the correct expressions, when we return an instance of FunctionExpression we lost
the mplementation of operator(), and deriv of the elementary functions.

Infact FunctionExpression::operator(), cannot access ANY implementation
of a class that derives from FunctionExpression.

Here the we have the following error.

FunctionExpression = elementary_function_nstance->devivate();
This automatically cast to elementary_function to FunctionExpression.
so we lost the behaviour of operator() and deriv